refactoring

+++ cppfand/FileD.h
@@ -32,7 +32,7 @@ public:
 
 	longint IRec = 0;
 	RdbPos ChptPos;           // zero for Rdb and FD translated from string 
-	WORD TxtPosUDLI = 0;      // =0 if not present; urcuje zacatek odstavcu #U #D #L #I
+	size_t TxtPosUDLI = 0;      // =0 if not present; urcuje zacatek odstavcu #U #D #L #I
 	FileD* OrigFD = nullptr;  // like orig. or nil
 	
 	WORD CatIRec = 0;
